Redoing scene
had some issues with scene.java where the image wouldnt' load proper so I decided to restart that from basically scratch to get it working properly. It hink it had to do with the way I was making the frame/panel/stuff? Not entirely sure though. Mostly working to a degree and able to animate it now as well. Just have to rewrite all the other code to fit back in with the new concepts.